// angular/projects/account/public/src/enums/components.ts
export const enum eAccountComponents {
  AuthWrapper = 'Account.AuthWrapper',
  Login = 'Account.LoginComponent',
  Register = 'Account.RegisterComponent',
  EmailConfirmation = 'Account.EmailConfirmationComponent',
  ForgotPassword = 'Account.ForgotPasswordComponent',
  ResetPassword = 'Account.ResetPasswordComponent',
  ManageProfile = 'Account.ManageProfileComponent',
  TenantBox = 'Account.TenantBoxComponent',
  ChangePassword = 'Account.ChangePasswordComponent',
  RefreshPassword = 'Account.RefreshPasswordComponent',
  PersonalSettings = 'Account.PersonalSettingsComponent',
  Logo = 'Account.LogoComponent',
  MySecurityLogs = 'Account.MySecurityLogs',
  ProfilePicture = 'Account.ProfilePicture',
  LinkLogged = 'Account.LinkLogged',
  SendSecurityCode = 'Account.SendSecurityCode',
}